Don't limit search_issues() results by default
The default maxResults=50 is a hazard that we ran in to
because if you're not explicitly specifying maxResults
you don't expect to have to handle paging.

Since the implementation neatly handles batch retrieval
of all results, it seems appropriate to use this logic
to get all search results by default unless you explicitly
limit this by setting maxResults.