Restore missing delay in ResilientClient and & implement exponentional backoff
This addresses two issues related to high concurrency/load.
1.) We weren't sleeping for the 502, 503, 504 statuses, resulting in immediate retry.
2.) Adds jitter to reduce thundering-herd effect that can happen when there is a sudden increase of load at high concurrency.

The current delay logic results in a max of 60 (10 + 20 + 30) seconds of delay before final success/failure.

The jittered delay in this patch will result in a max of 120 seconds of delay (20 + 40 + 60) but each delay is multiplied by a random number between 0 and 1 so the average delay is still 60 seconds. This implementation is the one described as "full jitter" at http://www.awsarchitectureblog.com/2015/03/backoff.html.

Also a minor formatting change in the log message for 5xx case. "Err: <Response [503]>" is now "Err: 503 Service Temporarily Unavailable". Logging the reason is helpful for debugging.