Fix setup.py's setup_requires and requirements.txt
This patch fixes an issue where installing `jira` using `pip`, or really
any use of the `setup.py` file, causes `pip` or `python` to hang, unless
you are able to contact PyPI.

Listed among `setup.py`'s clauses was `setup_requires`, which pointed to
`pytest`, leading `pytest` to be a setup requirement unnecessarily. This
patch removes that.

One workaround was  _manually_ `pip` installing `pytest` _before_
installing the packages listed in `requirements.txt`, but
`requirements.txt` is the canonical point of reference for
`pip`-installed packages.

Furthermore, `pytest` is not and should never be a build or production
requirement.  It is strictly a test requirement, and is thus listed in
`requirements-dev.txt`.

Required for installation, but absent from `requirements.txt`, was
`requests-oauthlib>=0.3.3`, so I've added that in this patch, too.

These fixes are particularly useful for those running out of clean
virtualenvs, or where you have your own PyPI set up, or on a network
without access to PyPI.