Improve auto-generated changelog
The earlier changelog was just a dump of the git log summary without any
markers due to which it was difficult to understand what changes were
introduced between the released versions.

This commit fixes it by generating the changelog using the
'gitchangelog' tool (https://github.com/vaab/gitchangelog). This tool is
added as a dev-dependency and is configurable via the '.gitchangelog.rc'
file. Also, the release script has been updated to use this tool.