Add support for custom session cookies to JIRA
This patch allows to connect to a JIRA server behind a single-sign-on
system that requires the client to pass a cookie for enabling HTTP
basic authentication.

The implementation has been done in a generic way by extending the
`options` parameter of the JIRA constructor to support a `cookies`
dict (similar to `headers`) which is passed to request's session
cookies.

Unfortunately just passing the cookie with a "Cookie" header is not
enough because the cookie has to remain also in all further HTTP
requests of the session.