Update client.py to fix the __init__ docstring
It appears that sphinx (as would make sense) does not attempt to
document dunder / _private_ methods. As a result i have moved the
information into the class documentation which I believe had yielded the
desired results. We can now see the parameters for the __init__ in the
documentation which should hopefully alleviate some misunderstandings

Ref #465