chore(github): adds jira instance type to bug-report template
Self-hosted and Atlassian's cloud hosted versions of Jira start to have diverging APIs. Hence it is relevant to know for a bug report on what type of JIRA instance this error occurred, such that later we can more effectively triage issues (or know which issues to ignore if we decide to drop support for either of them).