Make user search GDPR compliant (#927)
* Make user search GDPR compliant

The `username` field is deprecated and Jira is gradually removing it
from the cloud instances. This is the second time such changes break
our integrations; until now, our workaround consisted of first searching
for the user based on the email, and then use the account id from the
response for the other requests. But now we cannot search anymore for
users based on the email, and we need to use the `query` field. Even if
we just pass the same exact value we passed to `username`.

I can imagine there're several places in the code that would require of
being changed for being completely GDPR compliant, but I have no time
at the moment to fix all of them :(

Thanks for your work!

* Remove redundant check

The payload object was already initialized with the `username` value,
so there's no need to check whether is specified or not.

* Init payload iwth both username and query

`query` is a different parameter from `username` so it shouldn't replace
it actually. Tested with v3 API and it's working.

* Update docblock regarding Jira cloud parameters

Co-authored-by: adehad <26027314+adehad@users.noreply.github.com>

* Fix optional `user` param

* Update docblock

Co-authored-by: adehad <26027314+adehad@users.noreply.github.com>

* Update docblock

Co-authored-by: adehad <26027314+adehad@users.noreply.github.com>

* Update docblock

Co-authored-by: adehad <26027314+adehad@users.noreply.github.com>

Co-authored-by: adehad <26027314+adehad@users.noreply.github.com>