Allow disabling searching for user on assign_issue
This adds a parameter to the `_get_user_id()` method which allows for
selectively disabling the search for the user that said method performs.
This allows for code to perform it's own user search prior to calling
methods which use this method (e.g. `assign_issue`). This fixes an issue
where the search would return the wrong user due to the ordering of
users, and prevents the code from issuing duplicate search calls to the
Jira server when the user has already done so.

Closes #1284